centrifuge: Add migrations module
